+++ cloud-tenant-cd/OWNERS
@@ -1,0 +1,2 @@
+mortent
+bjorncs
