+++ hosted-api/src/main/java/ai/vespa/hosted/api/DefaultApiAuthenticator.java
@@ -1,10 +1,7 @@
 // Copyright 2020 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package ai.vespa.hosted.auth;
+package ai.vespa.hosted.api;
 
-import ai.vespa.hosted.api.ControllerHttpClient;
-import ai.vespa.hosted.api.Properties;
-
-public class ApiAuthenticator implements ai.vespa.hosted.api.ApiAuthenticator {
+public class DefaultApiAuthenticator implements ai.vespa.hosted.api.ApiAuthenticator {
 
     /** Returns a controller client using mTLS if a key and certificate pair is provided, or signed requests otherwise. */
     @Override
