+++ pom.xml
@@ -126,9 +126,9 @@
         <module>standalone-container</module>
         <module>statistics</module>
         <module>storage</module>
-        <module>tenant-auth</module>
         <module>tenant-base</module>
         <module>tenant-cd-api</module>
+        <module>tenant-cd-commons</module>
         <module>testutil</module>
         <module>vdslib</module>
         <module>vespaclient-core</module>
