--- tenant-auth/OWNERS
@@ -1,1 +1,0 @@
-jonmv
