--- tenant-auth/README.md
@@ -1,2 +1,0 @@
-<!-- Copyright 2020 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root. -->
-# Utilities that authenticate users to the hosted Vespa API, or to hosted Vespa applications.
