--- tenant-auth/src/test/java/ai/vespa/hosted/auth/AuthenticatorTest.java
@@ -1,6 +1,0 @@
-// Copyright 2020 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package ai.vespa.hosted.auth;
-
-public class AuthenticatorTest {
-
-}
