+++ tenant-cd-commons/OWNERS
@@ -1,0 +1,2 @@
+bjorncs
+mortent
