+++ tenant-cd-commons/README.md
@@ -1,0 +1,3 @@
+# tenant-cd-commons
+
+Contains shared implementations of APIs/interfaces of `tenant-cd-api`.
